requirements.txt
# Core dependencies
torch>=2.0.0
diffusers>=0.19.0
transformers>=4.29.0
accelerate>=0.20.0
safetensors>=0.3.1
numpy>=1.24.0
Pillow>=9.4.0
opencv-python>=4.7.0
scipy>=1.10.0

# LLM integration
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
langchain_community>=0.0.10
pydantic>=2.0.0

# Configuration and utilities
python-dotenv>=1.0.0
PyYAML>=6.0
tqdm>=4.65.0

# Web interface
streamlit>=1.22.0

# Development tools
ipykernel>=6.25.0

